+++ skfmm/base_marcher.h
@@ -18,7 +18,7 @@ extern "C" {
 class baseMarcher
 {
 public:
-  baseMarcher(double *phi,      double *dx,  int *flag,
+  baseMarcher(double *phi,      double *dx,  long *flag,
               double *distance, int ndim,    int *shape,
               bool self_test);
 
@@ -73,7 +73,7 @@ protected:
   double          * distance_; // return value modified in place
   double          * phi_;
   double          * dx_;
-  int             * flag_;
+  long            * flag_;
   int               error_;
   int               dim_;            // number of dimensions
   int               size_;           // flat size
