+++ skfmm/travel_time_marcher.h
@@ -4,7 +4,7 @@
 class travelTimeMarcher : public distanceMarcher
 {
 public:
-  travelTimeMarcher(double *phi,      double *dx, int *flag,
+  travelTimeMarcher(double *phi,      double *dx, long *flag,
                     double *distance, int ndim,   int *shape,
                     bool self_test,
                     double *speed) :
